+++ packages/frontend/src/features/queryRunner/semanticViewerPivotQueries.ts
@@ -121,6 +121,7 @@ export const getPivotQueryFunctionForSemanticViewer = (
             }, []);
 
         return {
+            queryUuid: undefined,
             results,
             indexColumn,
             valuesColumns,