+++ packages/frontend/src/hooks/dashboard/useDashboardChartReadyQuery.ts
@@ -114,6 +114,7 @@ export const useDashboardChartReadyQuery = (
             context,
             autoRefresh,
             hasADateDimension ? granularity : null,
+            invalidateCache,
         ],
         [
             chartQuery.data?.projectUuid,
@@ -126,6 +127,7 @@ export const useDashboardChartReadyQuery = (
             autoRefresh,
             hasADateDimension,
             granularity,
+            invalidateCache,
         ],
     );